add keyboard seeder

+++ keypedia/database/migrations/2022_01_16_151515_create_keyboard_table.php
@@ -20,7 +20,7 @@ class CreateKeyboardTable extends Migration
             $table->string('name');
             $table->longText('description');
             $table->string('imgPath');
-            $table->bigInteger('price');
+            $table->integer('price');
             $table->timestamps();
         });
     }
